pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xiaoan-parking-detection"
version = "1.6.0"
description = "共享单车智能停放检测系统"
readme = "docs/PROJECT_STRUCTURE.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "XiaoanNew Team"}
]
keywords = ["computer-vision", "object-detection", "yolov8", "vlm", "parking-compliance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "flask>=2.0.0",
    "openai>=1.0.0",
    "pillow>=9.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/XiaoanNew"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*", "app*", "config*", "utils*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = [
    "E402",   # module-import-not-at-top (sys.path 操作需要)
    "E501",   # line-too-long (已通过 line-length 控制)
    "E701",   # multiple-statements-on-one-line (紧凑条件语句允许)
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402"]
"app.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
